#60 fix(chat): 모델 분리
